+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-02T08:18:38.391Z",
+    "2026-04-02T09:20:24.795Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {