+++ packages/orq-rc/RELEASES.md
@@ -7979,3 +7979,13 @@ Based on:
 - [typescript v4.7.0-rc.41] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.41 - packages/orq-rc
+
+## 2026-04-02 09:23:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.42] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.42 - packages/orq-rc